requirements.txt
flask
google-generativeai
python-dotenv
mysql-connector-python
